break out console and control code to make it possible to create different views in the browser
