+++ style/common.css
@@ -1,0 +1,8 @@
+#container {
+    margin: 0 auto;
+}
+
+.twrap {
+    overflow: hidden;
+    margin: 4px 4px 4px 4px
+}
