+++ style/speaker.css
@@ -1,0 +1,47 @@
+.info {
+    background-color: black;
+    color: white;
+    font-family: "Lucida Console", Monaco, monospace;
+    width: 48%;
+    height: 80vh;
+    overflow: auto;
+    float: left;
+}
+
+.debug {
+    width: 100%;
+    clear: both
+}
+
+span.console_input {
+    display: block;
+    background-color: white;
+    color: black;
+    margin: 2px 2px 2px 2px;
+}
+
+span.heading {
+    color: Cyan;
+}
+
+span.description {
+    color: Lime;
+}
+
+span.next_step {
+    color: RebeccaPurple;
+}
+
+span.warning {
+    color: Red;
+}
+
+span.request_input {
+    color: Purple
+}
+
+.console_input {
+    width: 100%
+}
+
+
